chore: rename function file to match rule name
