+++ packages/ruleset/src/functions/major-version-in-path.js
