+++ packages/ruleset/src/rules/major-version-in-path.js
@@ -1,10 +1,10 @@
 /**
- * Copyright 2017 - 2023 IBM Corporation.
+ * Copyright 2017 - 2025 IBM Corporation.
  * SPDX-License-Identifier: Apache2.0
  */
 
 const { oas3 } = require('@stoplight/spectral-formats');
-const { checkMajorVersion } = require('../functions');
+const { majorVersionInPath } = require('../functions');
 
 module.exports = {
   description:
@@ -14,6 +14,6 @@ module.exports = {
   given: '$',
   severity: 'warn',
   then: {
-    function: checkMajorVersion,
+    function: majorVersionInPath,
   },
 };
